Removed the lock in the Message struct
Changed the exporting process method AddSetAndSendMsg to create
a new message and send it, rather than use a predefined message
object in exporting process struct.

This makes sure the message object is not shared by sendRefreshTemplates
go routine and the exporting process that is sending data sets.

This has to be extended when we support multiple sets in single
message and behavior may change a bit.