Revert "Modify Aggregation process methods (#158)"
Reverting this commit 7c0a0ae664f3720eee1d078598e7c825956f3df1.
Reason is the PR(vmware-tanzu/antrea#1949) that depends on this commit
is not yet ready. And to move the tag of go-ipfix in Antrea for other
PRs this commit should be present in the latest release tags. Therefore,
reverting this.