Add field AreExternalFieldsFilled for AggregationFlowRecord
Some fields like Pod labels will be filled externally on the IPFIX Mediator,
areExternalFieldsFilled is an indicator for IPFIX Mediator to check whether
these fields has been filled. This field is set to false when creating new
AggregationFlowRecord. Setting and utilizing this field is upto the user
and not used in go-ipfix library code.

Signed-off-by: Yongming Ding <dyongming@vmware.com>