Add field AreExternalFieldsFilled for AggregationFlowRecord (#211)
Some fields could be filled externally on aggregation records once before
exporting them in IPFIX mediator, e.g., metadata of a flow.
areExternalFieldsFilled is an indicator for IPFIX Mediator to check whether
these fields has been filled or not before exporting. This field is set to 
false when creating new AggregationFlowRecord. Setting and utilizing this 
field is upto the user and not used in go-ipfix library code.

Signed-off-by: Yongming Ding <dyongming@vmware.com>