Modify the basic structure `NewInfoElementWithValue`
Instead of relying on the fact that any data type can be interpreted
through interface{} type using their pointer, we move to InfoElementWithValue
interface. For each data type, we implement this interface with a separate struct.

This results in a very big file pkg/entities/ie_value.go
As there is no good support for generic data types in golang, we need to resort to
this.

In the process, we reduced allocations and memory consumed in the performance
test.