Create copies of net.IP and net.HardwareAddr when decoding IE values (#330)
When using the provided slice (provided by the collector) directly, the
garbage collector will not be able to release the underlying array while
references to the net objects exist. The underlying array is allocated
when reading the message from the TCP connection and can be 100s of
bytes in size. The collector will typically run aggregation and keep
references to the data record and its IEs for some time. During that
time, we cannot release the packet buffer, while new packets keep coming
in. To avoid the issue, we create a new slice, with a new underlying
array, and the packet buffer can be releases as soon as all IEs are
instantiated.

Signed-off-by: Antonin Bas <abas@vmware.com>