Retract version v0.10.0
The "target" of the v0.10.0 tag was "mutated" (the tag was created, then
deleted, then created again with a different target commit). This should
never be done for a Go module and can create issues for consumers of the
module, as the original version will be cached in module proxies, and
there will be a discrepancy between module proxies and the version
control repository (this can lead to checksum mismatch issues).

The best practice in that case seems to be to retract the invalid
version, and release a new one (we will release v0.10.1 in our case).

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>