Fix UDP template expiry in collector process
Prior to this change, for every template received by the UDP collector,
a new goroutine with a ticker was created, responsible for "expiring"
the template after the proper timeout. After expiration, the ticker was
stopped and te goroutine would complete (the ticker would only ever tick
once, so really a timer should have been used?). If a client were to
send a refreshed template, the old goroutine / ticker would not be
stopped, and the new template would be deleted by the old goroutine,
which is invalid. Additionally, starting a new goroutine every time a
template message is received is a bit expensive.

With the new implementation, we use a timer created with
time.AfterFunc. All the timers are handled by the go runtime with a
single goroutine, making this very lightweight. The timers are stored as
part of the template map. When a template is refreshed, we reset the
existing timer. When a template is deleted, we stop the timer.

This is the first in a series of patch to improve template management in
the collector process.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>