Delete existing template if a new invalid one is received (#383)
This is particularly useful for UDP collection, as there is no feedback
mechanism to let the sender know that the new template is invalid (while
with TCP, we can close the connection). If we keep the old template and
the sender sends data records which use the new template, we would try
to decode them according to the old template, which would cause
issues. Instead we will ignore data records for that observation domain
and template ID until a new valid template is received.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>