Manage templates per session instead of globally (#386)
As per the IPFIX specification, a template is specific to a transport
session. Observation domains are not guaranteed to be unique across
transport sessions.

When a transport session is closed (TCP connection closed or UDP session
timeout), all associated templates are removed. Note that prior to this
change, templates associated to a TCP connection were never cleaned up,
causing a memory leak.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>