Improve received stats in collector (#394)
GetNumRecordsReceived() was returning the number of IPFIX messages. This
was misleading as a message is different from a record, and a message
may include multiple records.

We add support for more stats (num messages, num template sets, num data
sets, num template records, num data records). At the moment, not all
of these stats are exposed through public methods.

GetNumRecordsReceived() is updated to return the number of received data
records. GetNumMessagesReceived() is introduced to return the number of
received IPFIX messages, which matches the previous implementation of
GetNumRecordsReceived().

Finally, we update the collector code so that stats are incremented
before sending the message on the channel, instead of after. This seems
more correct to me, but also it avoid a possible race condition in unit
tests, which assume that by the time a message has been received on the
channel (by the test code), stats have already been updated.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>