Prevent unnecessary memory allocations in WriteIPFIXMsgToBuffer
Calling GetBuffer on a record allocates a new buffer which is then
copied to the message buffer. This is not necessary as we can directly
encode the IE value to the message buffer.
The recommended way to do that is to call AvailableBuffer on the
bytes.Buffer object to get a []byte buffer with enough capacity, then
call Append-style methods to add data to the buffer (without allocating
new memory).

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>