Add Path MTU to Exporter inputs (#407)
It is not straightforward for the consumer of the library to determine
the maximum message size (for UDP) just based on the PMTU, as the amount
to deduct from the PMTU depends on the protocol used for the connection
to the exporter (IPv4 or IPv6). The protocol is only known after
resolving the collector address, which happens in InitExportingProcess.
As a workaround, we add PMTU to the exporter inputs, and let the
exporter compute the maximum message size. It is still possible to
request a specific maximum message size. If both are provided and the
exporter determines that the values are incompatible, an error will be
returned.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>