Add buffered exporter for IPFIX messages (#409)
SendDataRecords was added in #391 as a way to export data records more
efficiently. However, this method is not convenient to use for clients
which need to mutate the records after sending them (as is the case when
aggregation is used). To work around this issue, we introduce a buffered
exporter. The buffered exporter wraps the standard exporter
process. Whenever a data record is added, it is serialized to a byte
buffer immediately, and the buffer is sent when it reaches the maximum
message size.

Note that for convenience the buffered exporter also supports JSON
records, but in this case it just acts as a passthrough to the regular
("unbuffered") exporter process.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>