Bump Go to 1.24 for building
We also refactor the code so the Go version is read from a single place
(images/deps/go-version).
Note that unlike with previous Go version bumps, we do not update the go
directive in the go.mod file. This modules is meant as a library, and
updating the go directive too agressively means that all consumers of
the library have to also use a more recent version of Go. The go
directive should only be updated if we start using new language /
standard library features.

Finally, we upgrade to golangci-lint v2, which requires updating the
config file, as well as addressing some new lint errors.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>