Remove static certs to resolve test failures (#446)
CI has been broken because some static test certificates, which had been
generated with a 1000-day validity period, have just expired.

To avoid the issue altogether in the future, and to simplify the code,
we remove these static certificates, and instead generate them
on-the-fly when pkg/test/certs is imported.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>